Change `dontUseObjectIds` usage.
Instead of being a meta key, this is a model-level flag that indicates that the model doesn't use ObjectIDs for its primary key.  Waterline will use this to set a `collectionsNotUsingObjectIds` meta key that is an array of all collections that don't use ObjectID, so that both primary and foreign keys can be processed correctly.